Move undulator check out of common
