+++ src/mx_bluesky/common/plans/common_flyscan_xray_centre_plan.py
@@ -15,7 +15,6 @@ from bluesky.utils import MsgGenerator, make_decorator
 from dodal.devices.attenuator.attenuator import (
     ReadOnlyAttenuator,
 )
-from dodal.devices.dcm import DCM
 from dodal.devices.eiger import EigerDetector
 from dodal.devices.fast_grid_scan import (
     FastGridScanCommon,
@@ -23,7 +22,6 @@ from dodal.devices.fast_grid_scan import (
 )
 from dodal.devices.smargon import Smargon
 from dodal.devices.synchrotron import Synchrotron
-from dodal.devices.undulator import Undulator
 from dodal.devices.zebra.zebra import Zebra
 from dodal.devices.zocalo import ZocaloResults
 from dodal.devices.zocalo.zocalo_results import (
@@ -31,9 +29,6 @@ from dodal.devices.zocalo.zocalo_results import (
     ZOCALO_STAGE_GROUP,
     XrcResult,
     get_full_processing_results,
-)
-from dodal.plans.preprocessors.verify_undulator_gap import (
-    verify_undulator_gap_before_run_decorator,
 )
 
 from mx_bluesky.common.external_interaction.callbacks.xray_centre.ispyb_callback import (
@@ -80,8 +75,6 @@ class FlyScanEssentialDevices:
     zebra: Zebra
     zocalo: ZocaloResults
     smargon: Smargon
-    undulator: Undulator
-    dcm: DCM
 
 
 NullPlanType = Callable[[], MsgGenerator]
@@ -183,7 +176,6 @@ def common_flyscan_xray_centre(
 
     xrc_event_handler = XRayCentreEventHandler()
 
-    @verify_undulator_gap_before_run_decorator(composite)
     @bpp.subs_decorator(xrc_event_handler)
     def flyscan_and_fetch_results() -> MsgGenerator:
         yield from ispyb_activation_wrapper(
