+++ tests/unit_tests/hyperion/experiment_plans/test_hyperion_flyscan_xray_centre_plan.py
@@ -27,6 +27,7 @@ from mx_bluesky.common.parameters.constants import (
     DeviceSettingsConstants,
     PlanNameConstants,
 )
+from mx_bluesky.common.parameters.gridscan import SpecifiedThreeDGridScan
 from mx_bluesky.common.plans.common_flyscan_xray_centre_plan import (
     BeamlineSpecificFGSFeatures,
     FlyScanEssentialDevices,
@@ -378,3 +379,24 @@ class TestFlyscanXrayCentrePlan:
             and msg.obj.name == "attenuator"
             and msg.args == (1.0,),
         )
+
+    @patch(
+        "mx_bluesky.common.plans.common_flyscan_xray_centre_plan.run_gridscan_and_fetch_results",
+    )
+    @patch(
+        "dodal.plans.preprocessors.verify_undulator_gap.verify_undulator_gap",
+    )
+    def test_flyscan_xray_centre_does_undulator_check_before_collection(
+        self,
+        mock_verify_gap: MagicMock,
+        mock_plan: MagicMock,
+        RE: RunEngine,
+        test_fgs_params: SpecifiedThreeDGridScan,
+        fake_fgs_composite: FlyScanEssentialDevices,
+        beamline_specific: BeamlineSpecificFGSFeatures,
+    ):
+        mock_plan.side_effect = CompleteException
+        with pytest.raises(CompleteException):
+            RE(hyperion_flyscan_xray_centre(fake_fgs_composite, test_fgs_params))
+
+        mock_verify_gap.assert_called_once()
